requirements.txt
streamlit
pandas
pyswisseph
pytz
geopy 